Apply cutted corners containers to about and speakers
