+++ src/components/About.astro
@@ -5,8 +5,9 @@ import Diplodocus from 'src/assets/Diplodocus.astro'
 <section id='about' class='mx-auto my-8 max-w-screen-lg px-4'>
   <Diplodocus class='ml-auto mr-2 w-20 md:w-32' />
 
-  <div class='relative p-4 mb-8 text-black bg-white border-2 border-black shadow-[8px_8px_0px_0px_rgba(255,186,0,1)]'>
-      <p class='text-center'>
+  <div class='cut-corners-r mb-8 bg-black p-0.5'>
+    <div class='cut-corners-r bg-yellow pb-2 pr-2'>
+      <p class='border-b-2 border-r-2 border-black bg-white px-3 py-4 text-center text-black'>
         El DevFest Sucre es un evento tecnológico, organizado por la comunidad GDG Sucre, donde se
         comparten conocimientos sobre desarrollo de software, nuevas tecnologías y herramientas de
         Google.
