+++ src/components/speaker/ModalSpeaker.tsx
@@ -20,29 +20,17 @@ export const ModalSpeaker = ({ isOpen, onClose, data }: ModalPropsType) => {
 
   return (
     <div
-      className='pointer-events-auto fixed left-0 top-0 z-50 flex h-full w-full items-center justify-center bg-black/60'
+      className='pointer-events-auto fixed left-0 top-0 z-50 flex h-full w-full items-center justify-center bg-black/70'
       onClick={onClose}
     >
-      <div
-        className='max-w-xs bg-black p-0.5 dark:bg-white sm:max-w-xl md:max-w-full'
-        style={{
-          clipPath:
-            'polygon(calc(100% - 1rem) 0%, 100% 1rem, 100% 100%, 1rem 100%, 0 calc(100% - 1rem), 0 0'
-        }}
-      >
-        <div
-          className='bg-blue pb-4 pr-4'
-          style={{
-            clipPath:
-              'polygon(calc(100% - 1rem) 0%, 100% 1rem, 100% 100%, 1rem 100%, 0 calc(100% - 1rem), 0 0'
-          }}
-        >
+      <div className='cut-corners-r max-w-xs bg-black p-0.5 dark:bg-white sm:max-w-xl md:max-w-full'>
+        <div className='cut-corners-r bg-blue pb-2 pr-2'>
           <div
             className='flex max-h-screen w-full max-w-2xl flex-col border-b-2 border-r-2 bg-white dark:bg-black'
             onClick={e => e.stopPropagation()}
           >
             {/* Modal header */}
-            <div className='flex items-center justify-between border-b-2 border-black bg-blue px-1 py-2 text-white dark:border-white md:p-2'>
+            <div className='flex items-center justify-between border-b-2 border-black bg-blue px-1 py-1 text-white dark:border-white md:px-2'>
               <h3 className='ml-2 text-xl font-bold'>{data.name}</h3>
               <button
                 type='button'
@@ -60,7 +48,8 @@ export const ModalSpeaker = ({ isOpen, onClose, data }: ModalPropsType) => {
                   strokeLinecap='round'
                   strokeLinejoin='round'
                 >
-                  <path stroke='none' d='M0 0h24v24H0z' /> <line x1='18' y1='6' x2='6' y2='18' />{' '}
+                  <path stroke='none' d='M0 0h24v24H0z' />
+                  <line x1='18' y1='6' x2='6' y2='18' />
                   <line x1='6' y1='6' x2='18' y2='18' />
                 </svg>
                 <span className='sr-only'>Close modal</span>
@@ -69,18 +58,12 @@ export const ModalSpeaker = ({ isOpen, onClose, data }: ModalPropsType) => {
 
             {/* <!-- Modal body --> */}
             <div className='flex max-h-[70vh] flex-col space-y-4 overflow-scroll p-4'>
-              {data.descriptions && (
-                <>
-                  {data.descriptions.map((description: string, index: number) => (
-                    <p
-                      key={index}
-                      className='sm:text-md text-sm text-black dark:text-white md:text-xl'
-                    >
-                      {description}
-                    </p>
-                  ))}
-                </>
-              )}
+              {data.descriptions?.map((description: string, index: number) => (
+                <p key={index} className='sm:text-md text-sm text-black dark:text-white md:text-xl'>
+                  {description}
+                </p>
+              ))}
+
               <SocialIcons socialNetworks={data.socialNetworks} />
             </div>
           </div>
