+++ src/components/speaker/Speakers.astro
@@ -1,29 +1,17 @@
 ---
-// import SectionTitle from './SectionTitle.astro'
 import speakersData from '../../data/speakers.json'
 import SectionTitle from '@components/SectionTitle.astro'
-import { Speaker } from './Speakers'
+import { SpeakerCard } from './SpeakerCard.tsx'
 ---
 
 <div class='flex justify-center'>
-  <SectionTitle id='speakers' title='Speakers'/>
+  <SectionTitle id='speakers' title='Speakers' />
 </div>
 
 <div class='mx-auto my-8 max-w-screen-lg px-4'>
-  <div class='group grid grid-cols-2 gap-4 sm:grid-cols-2 md:grid-cols-3 pointer-events-none'>
-    {
-      speakersData.map(({ img, name, location, flag, descriptions, socialNetworks }, i) => (
-        <Speaker
-          id={i}
-          img={img}
-          name={name}
-          location={location}
-          flag={flag}
-          descriptions={descriptions}
-          socialNetworks={socialNetworks}
-          client:visible
-        />
-      ))
-    }
+  <div
+    class='group pointer-events-none grid grid-cols-2 gap-4 sm:grid-cols-2 md:grid-cols-3 md:gap-8'
+  >
+    {speakersData.map((data, i) => <SpeakerCard id={i} {...data} client:visible />)}
   </div>
 </div>
