--- src/components/speaker/style.css
@@ -1,7 +1,0 @@
-.max-scroll {
-    scrollbar-width: none; /* Firefox */
-  }
-  
-  .max-scroll::-webkit-scrollbar {
-    display: none; /* Safari y Chrome */
-  }
