+++ src/layouts/Layout.astro
@@ -67,6 +67,27 @@ import Header from '@components/Header.astro'
           scroll-behavior: auto;
         }
       }
+      :global(.cut-corners-r) {
+        clip-path: polygon(
+          calc(100% - 0.6rem) 0%,
+          100% 0.6rem,
+          100% 100%,
+          0.6rem 100%,
+          0 calc(100% - 0.6rem),
+          0 0
+        );
+      }
+
+      :global(.cut-corners-l) {
+        clip-path: polygon(
+          0.6rem 0,
+          100% 0,
+          100% calc(100% - 0.6rem),
+          calc(100% - 0.6rem) 100%,
+          0 100%,
+          0 0.6rem
+        );
+      }
     </style>
   </body>
 </html>
